ove_jupyter_utils/pyproject.toml
[build-system]
requires = ["hatchling>=1.5.0", "hatch-jupyter-builder >=0.8.3"]
build-backend = "hatchling.build"

[project]
name = "ove_jupyter_utils"
readme = "README.md"
license = { file = "LICENSE" }
version = "0.4.0"
description = "Utils package for ove-jupyter"
keywords = ["Open", "Visualisation", "Environment", "Jupyter", "OVE"]
requires-python = ">=3.8"
classifiers = [
    "Framework :: Jupyter",
    "License :: OSI Approved :: BSD License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
authors = [
    {name = "Brython Caley-Davies", email = "bc2918@ic.ac.uk"}
]

dependencies = [
    "python-dotenv==1.0.*",
    "markdown==3.4.*"
]

[project.urls]
Homepage = "https://github.com/dsi-icl/ove-jupyter"

[project.optional-dependencies]
test = [
    "coverage",
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
]

[tool.hatch.build.hooks.version]
path = "ove_jupyter_utils/_version.py"

[tool.hatch.build.hooks.jupyter-builder]
dependencies = ["hatch-jupyter-builder"]

[tool.jupyter-releaser.options]
version_cmd = "hatch version"

[tool.check-wheel-contents]
ignore = ["W002"]